Super sidebar: Test whether group items are serializable
This adds a shared test behavior for menus, to test whether they are
serializable as menu_items: `serializable as super_sidebar_menu_args`.

And another one for menus that are not serializable:
`not serializable as super_sidebar_menu_args`

Furthermore we enforce that serializable items should define an item_id.

And we have a minor fix and add the `active_route` to the Issues menu.