Skip package-and-test message on trivial MRs
The execution of package-and-test pipeline is required on merge request
targeting stable branches as a way to guarantee the changes introduced
are compliant from the Quality side. This is not necessary for
documentation changes or for merge request fixing flaky failures, so in
this case it's fine to skip the Danger message.

This commit updates the stable branch checks to skip the
package-and-test message on trivial merge requests with documentation
changes only or with fixes for flaky failures.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2839