New section to disable all 2FA
Created a new section to generate new recovery codes and disable all 2FA.
It is important to have a general section because it is not
only connected to TOTP but WebAuthn devices too.

This will become very important with the `webauthn_without_totp` feature
flag, because WebAuthn registration will generate recovery codes.

This is a MVC step for
https://gitlab.com/gitlab-org/gitlab/-/issues/378844#note_1212993074+

Changelog: changed