Rename "Configuration" to "Security configuration"
Within the "Security and Compliance" menu. We also make sure that all
the existing "Security Configuration" breadcrumbs match the sentence
casing. We also fix a few sentence casing issues in children pages:

- DAST Configuration => DAST configuration
- SAST Configuration => SAST configuration

The choice not to rename "API Fuzzing Configuration" to
"API fuzzing configuration" was made deliberately, because there are so
many wrong cased "Fuzz"es in the code base, that it would blow up the
scale of this changeset too much.

Part of: https://gitlab.com/gitlab-org/gitlab/-/issues/378546

Changelog: changed