Skip Danger failure when package-and-test is being created
The package-and-test execution is required for merge requests targeting
stable branches. Danger should fail when the package-and-test status is
canceled or manual, or when the pipeline is not present. Previous to
this commit there was a bug in the logic: Danger failed when the
package-and-test pipeline was being created, this is because it
only considered the downstream pipeline status rather than the bridge
status.

This commit ensures both status, the brige and the downstream pipeline
are considered, avoiding the Danger failure when the package-and-test is
being created.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2839