Added unit test for the `supported` function
The `supported` function is used to test if the browser supports the
credential management API.

We simplified the function using optional chaining.

We added unit testing for that function. There was some level of testing
of this function in
`spec/frontend/authentication/webauthn/authenticate_spec.js`. However,
this file would probably be deleted soon, when we replace the jQuery
application for a Vue component.