Hide edit for single user approval rules
Single user approval rules should be set to 1 always, as there is only 1
single user in the rule. How do you get 1 user to approve something
twice?

We hide the edit button instead of disabling it so users don't think
there should be a way to enable it that they are missing.