Remove 15.9 sanitize todo note
Removes note about making sure this migration runs. The reason is that
if we do todo sanitization, we should rather sanitize all existing
todos because there are probably some other edge cases when a todo is
not deleted. Or even better we might be able to avoid doing this todo
sanitization at all.

Either way, there is no reason to run this sanitization as it will be
superseded by different solution (or more general migration).