Validate deny requests setting using unsaved state
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/110599 introduced
a new `deny_all_requests_except_allowed` to application settings.

This change addresses the problem outlined in
https://gitlab.com/gitlab-org/gitlab/-/issues/377371#note_1271330392
that previously led the `ApplicationSetting` model to be saved in an
invalid state due to the new setting.

This change has `ApplicationSetting` refer to the current, potentially
unpersisted, state of its `deny_all_requests_except_allowed` property to
control whether the validation includes the
`deny_all_requests_except_allowed` setting or not.

This means an instance:

1. Cannot enable `deny_all_requests_except_allowed` while a URL would
   break the rule. The admin must first change that URL property, or add
   it to the allow list.
2. Can always disable `deny_all_requests_except_allowed`, even if
   somehow they are in a state where there are URL settings that breaks
   the rule.