Update @graphql-eslint-plugin to 3.16.1
Transform "operations" configuration key
for "graphql-eslint-plugin" into a string.
There is an issue with "graphql-eslint-plugin" cache implementation.
The cache is implemented as a Map which stores an array
as a reference. Having the "operations" key as an array invokes
that issue so we never hit cache because arrays are compared
by a reference in JavaScript. See below:

let arr = ['1'];
let cache = new Map();
cache.set(['1'], 1);
cache.has(['1']); // false