Reduce the chances of deadlocks for builds migrations
When we execute migrations that require locks in CI,
we usually start with `ci_builds` first and then
the other tables because `ci_builds` is the busiest one.
Creating the artifacts and the updating the job,
locks `ci_job_artifacts` and `ci_builds` which is
conflicting with the migrations that already have a lock
on `ci_builds`, making the migration to fail with deadlock
errors.

Changelog: fixed