Use AccessExclusive lock for adding FK to post-deploy migration
In https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8521, we
saw that the attempt to add a foreign key from `p_ci_builds_metadata`
to `ci_builds` failed due to a deadlock. The deadlock appears to be
happen because while `ShareUpdateExclusiveLock` allows reads but
prevents concurrent modifications to `ci_builds`, an open transaction
trying to upgrade to a `RowExclusiveLock` will deadlock while the
`ALTER TABLE` attempts to grab an `AccessExclusiveLock`. This can
happen on a busy system like GitLab.com.

`AccessExclusiveLock` is the heaviest lock that prevents any read
or writes from the table. In this case, we need to use this to
ensure we have complete control of the tables and prevent read
transactions from starting in the first place.