Convert user profile tabs to a sidebar panel
This creates a sidebar panel holding the user profile's tabs' links.
This is currently only visible with the super sidebar enabled as we are
redesigning this page as part of that initiative. The whole page is
currently being migrated to Vue, so we're doing the bare minimum to
support the navigation redesign effort for now. This will likely need to
be revisited later on once the Vue migration has progressed a little
more.