Revise Torsten's changes and fix problems
Swap out image for an anonymized one that is MUCH smaller in file
size and display size. Rearrange various paragraphs on the page
to make the image flow with the text. Create a new subheading
on the page to explain how to see Code Owners for a file. Bump
the video down so it's not so early in the page. Fix spelling
errors.