Fix unlocking of job artifacts on pipeline state transitions
Job artifacts are now unlocked after each pipeline transition to:
- success
- failed
- manual (blocked)

When the last pipeline on the ref is not successful,
the unlock process will keep the job artifacts from the
last successful pipeline as well as the last pipeline.
The reason is that the last pipeline (failed or blocked),
could eventually succeed and become the last successful pipeline
that should have the job artifacts locked.

Changelog: fixed
Co-authored-by: Sam Kim <skim@gitlab.com>