Fix batch loader for CI project includes
We should not use a global changable method inside a batch-loader block.
The mistake was to use `project` inside `can_access_local_content?` and
`sha`. So, it was "cached" for other includes.

In this commit, we are using "project" as the `for` parameter of
batch-loaders. One downside of this is that we still have N+1 queries
for projects.

In this commit, we are also changing the tests to cover this case.

These changes are behind the FF ci_batch_project_includes_context