Add link to Zoekt query syntax from exact search docs
In https://github.com/sourcegraph/zoekt/pull/539 we added a lightweight
placeholder for Zoekt query syntax documentation in Zoekt. We ultimately
want Zoekt to remain the source of truth for the query syntax but the
docs are very light on at the moment so this page linking into some
difficult to read HTML is just a start. We have an in principal
agreement in
https://github.com/sourcegraph/zoekt/pull/539#issuecomment-1453864835 to
make this markdown doc more comprehensive but for now we're at least
linking people up so that I have some way to find out how to do
searches.