Update documentation to match Gitaly's new configuration structure
Omnibus previously had a custom configuration structure for Gitaly
that didn't match the actual structure of Gitaly. This made a number
of things more difficult than they had to be. Omnibus was recently
updated to place Gitaly's configuration under gitaly['configuration']
hash which gets directly marshaled as TOML for the final Gitaly
configuration. While there is still backwards compatibility handling
in place, we should update the documentation to use the new structure
in preparation for deprecating the old configuration keys. This commit
updates Gitaly's documentation to refer to the new keys.