Pass through `RUBY_VERSION` to release-environments child pipeline
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2674

We can see when attempting to start jobs in the release environments
pipeline (example https://gitlab.com/gitlab-org/gitlab/-/jobs/3940694610)
that we get a bad image error due to `RUBY_VERSION` not being passed
as a variable to the downstream pipeline.

This change corrects this.