Remove index on events.action
This index has
[been unused for at least 8 weeks according to Prometheus](https://thanos-query.ops.gitlab.net/graph?g0.expr=sum(rate(pg_stat_user_indexes_idx_scan%7Benv%3D%22gprd%22%2C%20instance%3D~%22patroni-main-2004.*%22%2Cindexrelname%3D%22index_events_on_action%22%7D%5B1w%5D))%20by%20(indexrelname)&g0.tab=0&g0.stacked=0&g0.range_input=8w&g0.max_source_resolution=0s&g0.deduplicate=1&g0.partial_response=0&g0.store_matches=%5B%5D).

Conceptually I also can't see a good use case for this index. It would
only be useful if we had an admin page which allowed filtering through
all user activity and I couldn't find any such thing. Or it might be
useful if there was an admin API that allowed filtering all events but
from what I can see the
[Events API](https://docs.gitlab.com/ee/api/events.html) is always
scoped to a project or user.

Lastly I tried to look around the codebase where this could be used. We
do filter by `where(action: ` in some places like `EventsFinder` but in
all usages I see that we're always scoped to something more selective
than the action which would mean the action index would never be used.
This is all supported by metrics and since this is the largest index
which is unused in our system it makes sense to delete it.

Changelog: performance