Log slowest HTTP requests
This would help identify if certain HTTP requests are slow, or whether
there's a general problem with scheduling them. This commit adds a
`external_http_slow_requests` field to the log payload that will
contain an array of the top slowest requests with the hostname, port,
method and path.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/364561

Changelog: added