Address polling edge cases in runner registration
This change addresses two cases which may happen during the display of
the runner registration page.

- After the token is no longer available (1 hour) the runner token
disappears from the screen, leaving the user unable to copy the token.
- If the runner is deleted (in another screen) the page stops updating
the runner information.

While currently the 1 hour time seems long, it may be shorted to a few
minutes as the feature is refined.

Changelog: fixed