Fix to prevent unlimited CI includes
This MR fixes the bug that allowed unlimited nested includes
in the CI config by enabling the feature flag ci_fix_max_includes.
The fix involves re-ordering the operations
in the file expansion process such that all included files are
correctly added to the structure that we use to obtain the
total file count.

Changelog: fixed