Major structural revision of 'branches' page
Does most of the long LONG overdue structural tidying that will
allow us to add many of the basic branch-related functions to the
page. It also creates a scaffolding to let the new branch-protections
page fit neatly into the page order, rather than hiding it down at
the bottom of the page.