Fix inconsistent behaviour in work items comments
The main issue is related to the use of the shortcut ctrl/cmd + enter
while saving the comment. As the comment is stored in the local
storage, it was getting cleared on form submission but restored
as the key input event was calling setCommentText method. Clear the
local storage when update mutation is successfully completed.

Changelog: fixed