Set KAS connect-src CSP on demand
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/104504 we've
introduced a `connect-src` directive for the KAS subdomain (in case it
is on a subdomain) *globally*. However, this is not really necessary and
an on-demand approach is preferred.

*on-demand* meaning that a controller can include the `KasCookie`
concern to configure the correct CSP so that the KAS cookie can actually
be used. The KAS cookie doesn't make any sense without the CSP config,
so it naturally makes sense for it to be part of the `KasCookie`
concern.