Stop triggering subscription when MR marked as checking
Since there is no difference in the UI between `unchecked` and
`checking` merge status, there's no use to trigger the subscription
when MR is marked as checking. We only allow to transition a MR
to checking when it is from unchecked so the UI state before the
MR gets marked as checking is still valid.

This removes the `checking` and `cannot_be_merged_rechecking`
state from the `after_transition` block that triggers the
`mergeRequestMergeStatusUpdated` subscription.