Allow pipelines to be filtered by name in public pipelines API
Add an optional name parameter to the project pipelines public API
request, and include the names of pipelines in the response.

We have created a new entity called PipelineBasicWithMetadata since
PipelineBasic is used in a lot of places. With a new entity, we don't
need to change all of them at once.