Log merge_action_status when merging MR
To help debugging a QA test failure (setting MR to MWPS), it's good
to know what status is returned when `#merge!` is called in the
controller. This way, we can isolate the cause of the flaky failure.

To do that, we push to `Gitlab::ApplicationContext` so it can be
added to logs for `Projects::MergeRequestsController#merge` action
requests as `meta.merge_action_status`.