Fix nil and N+1 problems in max access level preloader
This preloader could use subqueries for the conditions on project_id and
user_id. There were two problems with this:

1. It did not save queries, because both the projects and users were
   expected to get preloaded anyway.

2. It could fail if the projects subquery loaded more projects than the
   eager loaded projects.

It also only preloaded permissions for users that had some level of
access to the requested projects, leaving an N+1 for users with no
access.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/395965