Allow WIT changes and fix issue types changes
This change introduces the ability to change any work item type to any
other by validating the hierarchy restrictions.

It also fixes an issue with existing issue type changes when they
are part of an epic. Until now it was possible to change the type of an
Issue to an Incident, even though the Issue was part of an Epic.
This should be prevented as Incidents do not support Epics.

Changelog: added