Remove long-running transaction while deduplicating the findings
This change removes a problematic transaction while updating the
findings of the scan as deduplicated. To inform the user if the logic
fails, we now add a processing error to scan and set it as failed.

If this logic fails, findings of the related security scan will not be
ingested as part of the `IngestReportService` logic.

Changelog: fixed
EE: true