Revert "Gitlab::UrlBlocker: Ensure absence of URI scheme with `schemes: :none`"
This reverts commit 77870fca5380b0a149f287b16e87702d27b727da.

According to https://www.ietf.org/rfc/rfc3986.txt URIs without a scheme
are invalid. Weirdly, they are accepted by `Addressable::URI.parse`.

For example:
  uri = Addressable::URI.parse("192.168.1.2")
  uri.scheme   # => nil
  uri.hostname # => nil
  uri.path     # => "192.168.1.2

With this it doesn't make sense to enforce URI without a scheme when
passing `schemes: :none` in UrlBlocker as it won't be able to resolve
those URIs because `hostname` is `nil`:

  Gitlab::UrlBlocker.validate!("192.168.1.2", schemes: :none)
    # Gitlab::UrlBlocker::BlockedUrlError: Host cannot be resolved or invalid
    # Caused by SocketError: getaddrinfo: Name or service not known