Add feature flag check to GraphQL trigger
When `realtime_mr_status_change` feature flag is enabled, the
frontend will subscribe to the `mergeRequestMergeStatusUpdated`
subscription.

If something goes wrong during rollout and the frontend is still
connected even after disabling the feature flag, backend will still
trigger the subscription and push updates to the frontend.

To avoid that, this adds the same feature flag check to the
GraphQL trigger. This way, when the feature flag gets disabled,
backend won't trigger the subscription as well.