Update gettext-extractor-vue and dependencies
This updates gettext-extractor-vue and related dependencies. We also
make tiny changes to our gettext extraction, by adding a simple `guard`
to the Vue Extractor. The Extractor will now skip files that do not at
least contain a `__` which is part of all our gettext function names in
the frontend.

We are "only" saving half a second on my machine, but on slower
machines, we might save a little more.