Fix `Remember me` omniauth
The ID of the `Remember me` checkbox element was duplicated. The same
ID, `remember_me`, was used for the main login (upper) and the omniauth
logins (lower portion). This created a couple of undesired side-effects:

1. Selecting on the lower `Remember me` checkbox, checked the upper
   checkbox (main login).
2. More importantly, the `Remember me` was broken because it didn't add
   the parameter `remember_me=1` to the `action` path of the form
   surrounding the ommiauth logins.

This commit fixes the issue by replacing the `remember_me` ID to a more
specific name, `remember_me_omniauth`, for the omniauth logins.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/385419