Improves dataframe rendering in Python Notebooks
Due to security concerns, all rendering of html embedded in Jupyter
Notebooks became sandboxed in an iframe, causing poor user experience
for users.

We now identify if the html is a dataframe, and display it accordingly.
It also adds interactivity to the dataframe, such as sorting and
filtering the data.

Changelog: fixed