Extremely light edits to keep work moving
I reshaped and shortened the initial bulleted list to keep the
language tight and parallel. However, I'm opting to stay out of the
rest of the work because I want to let Denys build as he sees fit.
I can edit after the fact. Anything beats what we have now.