Don't generate invalid SQL checking foreign keys
When migrating many versions of gitlab at once, it was possible to call
foreign_key_exists? with a schema cache too old for rails to know about
the constrained_columns column of the postgres_foreign_keys view.

This manifested with errors like
`PG::InvalidTextRepresentation: ERROR: malformed array literal`

Reload the schema cache in this case to avoid the error.

Changelog: fixed