Fix N+1 query for max access level in project lists
This removes the feature flag
`fix_users_helper_load_max_project_member_accesses` that was used to guard
the rollout on GitLab.com because this is the landing page for logged in
users.

Part of https://gitlab.com/gitlab-org/gitlab/-/issues/214037

See rollout issue https://gitlab.com/gitlab-org/gitlab/-/issues/396647
for the details on the feature flag's rollout.

Changelog: fixed