Allow HEAD ref when viewing blob, commits and compare page
When viewing blob, commits, and compare page, we validate the ref/s
being passed via params if they're valid refs.

While `HEAD` is not a valid branch name, it's a representation of
whatever the default branch is (e.g. `main` or `master` usually).

To allow viewing blobs, commits, and compare page for the `HEAD`
ref, a `skip_head_ref_check` flag is added to
`Gitlab::GitRefValidator#validate`. This will skip the `HEAD`
ref check if set to `true`. It's set to `false` by default.

Changelog: fixed