Hide inactive approvers
When a user becomes inactive (e.g. blocked, banned, deactivated),
they can't approve a MR. But as of this writing, we still show
blocked users when they're added as approvers to an approval rule.

To prevent that, we filter inactive users from the list of
approvers.

Changelog: fixed
EE: true