Add Show More button to limit shown avatars
Part of the design is limiting the number of avatars for a deployment or
approval rule, and this is a common pattern used in the GitLab UI. So
much so that is should likely be shared. However, I want to sit on it a
bit. It currently has no way of requesting "even more", and I'm not sure
it should.

So, the beginning of a re-usable shared Show More simply
limits the shown list to the number of items. Clicking on the button
displays "Show less", but I am open to word changes.