Display fallback emoji images with correct aspect ratio
The emoji are natively square (64px x 64px).  By displaying them
20px x 16px we squash the images vertically.

Because the line height we use is 16px we cannot display the emoji
to match the current width of 20px as this will force any line
containing an emoji to display too high.  So scale to 16px x 16px.

After this change the CSS for img.emoji is redundant so this is also
removed.

Changelog: fixed