Revises emoji section of Markdown page
As a prototype for other potential changes to this page, this
commit changes the emoji display from sequential to tab-based.
It also adds a section about GitLab custom emoji, because this is
an undocumented feature that most folks don't realize exists because
it's not in the emoji picker yet.

It also adds some extra CSS formatting to the images to make them
display better.