Update server hooks documentation
With GitLab 15.11 a new Gitaly CLI command is being introduced to set
custom git hooks for a repository. The old method of setting the git
hooks directly on the file system still no longer be supported. This
change updates the documentation to explain how to use this new command
to set hooks for a repository.