Add new super sidebar menus for projects
As of now these menus are dummy menus which are mostly not _rendered_,
as they have no menu items. The reason why we introduce them all within
this commit: It will make it easier to have parallel MRs which actually
utilize the new menus, resulting in less conflicts (e.g. translations,
etc.)