Super sidebar: Replace a placeholder when adding menus
In the first iteration of our super sidebar menus, we used a property on
the menu items in order to control where they are inserted. The
property `super_sidebar_before: :foo` is used to insert the current menu
item before the menu item `:foo`.

In practice this has proven to be a little difficult for a few reasons:

1. The CE/EE split makes it hard to predict where exactly an item will
   be, because depending on CE / EE the order of items might be
   different, as the EE version might inject items somewhere in the
   middle.
2. In the Super Sidebar we are pulling items from many different other
   menus into another menu. Now this also depends on the order these
   menus have been added to the original Panel
3. A recent change in our requirements, also wants us to reorder a few
   elements inside the panels and I assume that we want to reorder items
   more easily in the future.

So the idea is that for the Super Sidebar we just create an array of
`NilMenuItem` and let them have the same `item_id` as the real items.
These placeholders are then going to be replaced with the real item
during the `transform_old_menus` conversion. In case no placeholder was
defined, the real item is just appended to the end of the list.

This technique will render `super_sidebar_before` obsolete, and the plan
is to remove it once all menus are using the new technique.