Project super sidebar: Analyze category
This moves all Analytics menu items into the new Analyze category
See: https://gitlab.com/gitlab-org/gitlab/-/issues/393190. We should
note that the items still need to be renamed in the super sidebar, but
we are mainly dealing with the order first.

We also rename the following items in the super sidebar only:
 - CI/CD => CI/CD analytics
 - Repository => Repository analytics
 - Value stream => Value stream analytics
 - Dashboards => Application analytics
 - Merge request => Merge request analytics
 - Issue => Issue analytics
 - Code review => Code review analytics