Append tag notes when creating a new tag too
Releases can append an annotated tag's notes on creation, but only if
the annotated tag already exists.

A user can also create an annotated tag while creating a release, but
those notes would not be appended.

We correct that here by seeing if we are creating a tag while creating a
release, and then appending any added notes for that tag to the release
notes.

Changelog: fixed