Add new super sidebar menus for groups
As of now these menus are dummy menus which are _mostly_ not rendered,
as they have no menu items. We will fill them with menu items in
follow-ups. The reason why we introduce them all within
this commit: It will make it easier to have parallel MRs which actually
utilize the new menus, resulting in less conflicts.