Group super sidebar: Monitor category
This moves all Observability menu items into the new Monitor category.
See: https://gitlab.com/gitlab-org/gitlab/-/issues/393190

Note that we have added the Monitor `super_sidebar_parent` to a few
items which are not rendered at all. This should not be problematic
though.

Furthermore, if you try and navigate to one of the Monitor items, the
layout will switch back to the old layout. This will be addressed with:
https://gitlab.com/gitlab-org/gitlab/-/issues/403072