Add workItemConvert mutation
This MR introduces a mutation to change the work item type.

It's using a separate mutation, instead of the existing update mutation,
to prevent cases where a new type gets set + new data for a widget that
doesn't exist in the new type.

When a widget does not exist in the new work item type, widget data
will be deleted.

Changelog: added