Use "job" instead of "processable" and "status" in pipeline processing
In https://gitlab.com/groups/gitlab-org/-/epics/9991, we aim to convert
all CommitStatus related models to Ci:Job models. With that epic,
we aim to replace `statuses` with `jobs`.

As a part of the refactoring pipeline processing effort, this commit
will contribute to it by trying to achieve the "jobs" goal.