Skip epic cache update if epic can not be saved
When deleting a group hierarchy, it's possible that cache counts for an
epic are updated at the same time when epic's group is deleted (which
causes cascade deletion of the epic itself). In that case epic couldn't
be saved in DB because validation fails (as referenced group doesn't
exist anymore). We can just skip cache update if epic can not be saved
anyway.

Changelog: fixed
EE: true