Add an action to parse the note hash post-load
Now that the UI has totally asynchronous load of ALL of the UI data,
we need to trigger parsing that used to only happen once.

That is:

- Discussions are loaded async
- Files are loaded async
- Any number of files may be present, from 0—total.

Reading the note hash could extract a note ID that doesn't yet exist
in the `notes` store. Or, the existing note may point to a file that
hasn't yet loaded.

For the former, we will simply trigger this new action (which extracts the
URL hash) when the notes indicate they have updated.

For the latter, we will trigger `fetchFileByFile` to ensure that the file
that this note points to is actually loaded. `fetchFileByFile` will
quickly short-circuit if it has already been loaded, so there's no
harm in triggering it.