[Geo] Fix token timeout for container registry sync
We use memoization for an HTTP client when it's set up
with auth data and other parameters.
The token is relatively short-living, 5 minutes by default,
so for large containers, we can get a timeout due to
an authentication error.
This MR adds a memoization utility with an expiration time
and uses it for registry clients.

Changelog: fixed