Performance: Preload pipeline in PipelinesController to fix N+1
It appears that when `merge_trains` feature is used we are
fetching the pipeline and merge request object over and over.

Even though the N+1 is solved by cached query, the AR object initialization
makes this endpoint to spend 6s instead of 100ms on GitLab.com
for GitLab Rails pipelines.

Changelog: performance