Move :aggregate_failures into test cases
If it's put in shared_examples or shared_context, the shared_examples
or shared_context will be included into the test cases where the same
tag is used, and that's definitely not what we want.

This is discovered at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116298#note_1335497451

In this merge request, we push :aggregate_failures into the test cases
where it's needed.