Update gitlab-mail_room to v0.0.23
This version is already used by GitLab Omnibus and Cloud Native
GitLab, but we had to avoid updating this dependency in GitLab Rails
due to the need to depend on an updated net-imap gem
(https://docs.gitlab.com/ee/development/emails.html#rationale). Now
that we've updated to Ruby 3.0.5, we can safely bump this version.
This version bump is also needed for Ruby 3.1 compatibility.

Changelog: changed