Send queries to fetch schema cache to replicas
We have a CPU load issue when pods are created. Many pods all request
schema information for tables either on boot, or on the first requests.
All of these queries go to the primary, and are then cached in the
schema cache.

It turns out that the schema queries always use the primary connection
pool. That is the default behaviour, see ConnectionProxy#method_missing.

To alleviate this load issue, we configure schema_cache to use replica
connection pools when possible.

Changelog: changed