Fix a replica latency bug on the Maven package registry upload endpoint
Maven package uploads happen in quick succession. Between those
requests, the package object must be created and re-used. Due to
database replica lag, we could situations where it was not re-used and
that would lead to rejecting uploads with either:
* 404 Not found
* 400 Bad Request, Name already taken
* 409 Conflict

This change completely fix the issue for 404 and 400. It also helps with
409 responses.

Changelog: fixed