Add /admin/search/zoekt APIs for controlling Zoekt rollout
As part of our rollout of Zoekt we're going to be enabling it 1
namespace at a time. This means that we'll need tooling to simplify this
process. Originally I was planning on doing this via the Rails console
but I've since learnt any rails console commands require a C2 change
request. This requires lots of approvals and someone with access to
perform the steps. This seemed quite wasteful so I've subsequently
realised this would be better to roll out via chatops. Chatops requires
an API to execute the steps so this adds some admin APIs we'd need to
run from chatops.

Also these same APIs might end up being useful for other administrators
managing Zoekt for their instances or possibly to support a future UI.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116650
EE: true