Add CurrentUserTodos widget to add and mark todos
Adds a todo widget for updating Work Items and being able to add todos
or marking existing todos as done.
The latter allows to mark a specific todo of the work item to be done or
all todos of the work item for the user to be done.

Changelog: added