Clarify each_batch docs do not iterate non-unique column
Previously I was confused by this wording and got the impression that it
would just lead to inconsistent batch sizes. Given we know that it can
just get stuck in an infinite loop (which is actually documented further
down the page already) then the wording "proceed with extra caution" is
not as helpful as just telling you not to do it. They can already just
read on to the next sections about how they can iterate over non-unique
columns.