New nav: Fix syntax highlighting on Compare View
Similar to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/112605
where we fixed the syntax highlighting in the Commit View, this fixes
the syntax highlighting on the compare view.

However we are also changing the approach of how we calculate the "top"
of the sticky element. Rather than doing some calculation based on the
height of the navbar (and the existence of perfbar), we use the
`paddingTop` property CSS value of the `.layout-page` element. The
reason behind this: We put a lot of effort in setting the padding
correctly.