Allow CI job to need an undefined, optional job
Prior to this fix, you could not use `needs:` to create a job dependency
that points to a job added with `include:local:rules`. It would throw a
validation error because it does not see the needed job as defined.
This MR fixes the YamlProcessor (CI Linter component) so that
a job that needs an undefined, optional job will pass validation.

Changelog: fixed