Fix SD blob path when scan is run in "no git" mode
Checks whether the commit.sha value from the Secrets Detection
finding is the dummy default SHA before using it to compute the
blob path.

The dummy value of "0000000" is used when the `secrets` analyser
is executed in "no git" mode, and shouldn't be used to construct
the blob path because it won't exist in the repository.

EE: true
Changelog: fixed