GraphQL triggers: use symbols not strings
It's confusing that we used camel case strings
to reference triggers, when the actual trigger
field uses a snake-case symbol. While both uses
are correct, the official docs also use symbols
since they should match the field name used
for this trigger.

This has no impact on behavior, it's cosmetic.