Add Panel Type to Navbar click event
This adds the panel type (e.g. `your_work`, `group`) as an EventProperty
to our structured snowplow events. This will make it easier to group and
filter events by panel. The EventProperty is prefixed with `nav_panel_`,
so e.g. `nav_panel_your_work` or `nav_panel_group`.

We also have a fallback `nav_panel_unknown`, in case the panel is not
known.